SCWDEVOPS-3: Moved IO to command level.
This is an attempt to make the end effects of any command more transparent.

In this case I think it showed that commands like IngestAll could be abstracted to use th eindividual commands better.

It also showed that the collector would benefit from a pipeline between the gitlab reader and the processor.